+++ app/ui/settings/__init__.py
@@ -1,0 +1,1 @@
+"""设置窗口子包：按职责拆分自 settings_dialog.py。"""
